src/requirements.txt
Flask==2.3.3
Pillow>=10.4.0
piexif==1.1.3
Werkzeug==2.3.7
scikit-learn>=1.3.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
python-dotenv>=1.0.0
APScheduler>=3.10.0